replace routers param with a much simpler routePrefix param
