+++ lib/mapRoutes.js
@@ -9,15 +9,14 @@ const toobusy = require('toobusy-js')
 
 module.exports = app => {
   const express = app.get('express')
+  const router = app.get('router')
   const appName = app.get('appName')
   const logger = app.get('logger')
   const fsr = require('./tools/fsr')(app)
   const params = app.get('params')
+  const prefix = params.routePrefix
   const publicDir = params.publicFolder
-  let routers = params.routers
   let controllerFiles
-  let publicFolderDirs
-  let schemaLocation
 
   // ensure 404 page exists
   params.errorPages.notFound = path.join(app.get('controllersPath'), params.errorPages.notFound)
@@ -111,227 +110,53 @@ module.exports = app => {
   fsr.ensureDirSync(app.get('viewsPath'))
   fsr.ensureDirSync(app.get('controllersPath'))
 
-  // function to validate router parameters passed in
-  const checkRouters = function (customRouters, directory) {
-    let errorBool = false
-    const filesOrDirs = directory === 'controllers' ? 'files' : 'dirs'
-    customRouters = customRouters.filter(router => {
-      if (Object.keys(router).length === 0 && router.constructor === Object) {
-        return false
-      } else if (typeof router !== 'object' || typeof router.prefix !== 'string' || !/^[a-zA-Z0-9/_-]*$/.test(router.prefix) || !Array.isArray(router[filesOrDirs]) || router[filesOrDirs].length === 0) {
-        errorBool = true
-        return false
-      } else {
-        return true
-      }
-    }).map(router => {
-      router.prefix = router.prefix.charAt(0) !== '/' ? [router.prefix.slice(0, 0), '/', router.prefix.slice(0)].join('') : router.prefix
-      router[filesOrDirs] = router[filesOrDirs].filter(file => {
-        if (typeof file !== 'string') {
-          errorBool = true
-          return false
-        }
-        if (!fs.existsSync(path.join(app.get(directory === 'controllers' ? 'controllersPath' : 'publicFolder'), file))) {
-          logger.warn(`${app.get('appName')} failed to load the ${path.extname(file) === '.js' ? 'file' : 'directory'}: "${file}" to use with the router associated with ${directory === 'controllers' ? 'controller' : 'public'} prefix: ${router.prefix}`)
-          return false
-        }
-        return true
-      })
-      return router
-    })
-    if (errorBool) {
-      logger.warn(`Invalid configuration found in the "routers.${directory}" parameter. Please make sure it is coded correctly. See documentation at http://github.com/rooseveltframework/roosevelt for examples.`)
-    }
-    return customRouters
-  }
-
-  // validate routers parameter
-  if (Object.keys(routers).length > 0 && routers.constructor === Object) {
-    // check for a controllers schema
-    if (typeof routers.controllers === 'string') {
-      if (fs.existsSync(path.join(app.get('controllersPath'), routers.controllers))) {
-        schemaLocation = path.join(app.get('controllersPath'), routers.controllers)
-        routers.controllers = require(schemaLocation)
-      } else {
-        logger.warn(`Failed to load file: "${routers.controllers}". All controllers will be routed through the app level router.`)
-        routers.controllers = false
-      }
-    }
-    // validate supplied routing params
-    routers.controllers = Array.isArray(routers.controllers) ? checkRouters(routers.controllers, 'controllers') : false
-    routers.public = Array.isArray(routers.public) ? checkRouters(routers.public, 'public') : false
-  } else {
-    routers = false
-  }
-
-  // map public files for developer mode
+  // map statics for developer mode
   if (params.alwaysHostPublic || app.get('env') === 'development') {
-    // build list of public files
-    if (fsr.fileExists(path.normalize(app.get('publicFolder')))) {
-      // get newly created public symlinks
-      publicFolderDirs = klawSync(path.normalize(app.get('publicFolder')), { nofile: true })
-      publicFolderDirs = publicFolderDirs.map(publicDir => publicDir.path)
-      if (publicFolderDirs) {
-        let publicAssigned
-        let publicDirs
-        if (routers.public) {
-          publicFolderDirs.forEach(publicDir => {
-            publicAssigned = false
-            if (routers.public) {
-              routers.public.forEach(publicConfig => {
-                // remove duplicates from public dirs
-                publicDirs = [...new Set(publicConfig.dirs)]
-                publicDirs.forEach(definedPublicDir => {
-                  if (publicDir.includes(path.join(app.get('publicFolder'), definedPublicDir))) {
-                    app.use(path.join(publicConfig.prefix, definedPublicDir).replace(/\\/g, '/'), express.static(publicDir))
-                    publicAssigned = true
-                  }
-                })
-              })
-            }
-            // public wasn't assigned to a route so assign it to "/" + directory name
-            if (!publicAssigned) {
-              app.use(publicDir.split(app.get('publicFolder'))[1].replace(/\\/g, '/'), express.static(publicDir))
-            }
-          })
-        } else {
-          app.use('/', express.static(app.get('publicFolder')))
-        }
-      }
-    }
+    app.use(prefix || '/', express.static(app.get('publicFolder')))
   }
 
   // build list of controller files
   if (fsr.fileExists(path.normalize(app.get('controllersPath')))) {
     try {
-      // get controller files
-      controllerFiles = klawSync(path.normalize(app.get('controllersPath')), { nodir: true })
-      controllerFiles = controllerFiles.map(controller => controller.path)
-
-      // if a schema was supplied then splice that from the controller files
-      if (schemaLocation) {
-        controllerFiles.splice(controllerFiles.indexOf(schemaLocation), 1)
-      }
+      controllerFiles = klawSync(path.normalize(app.get('controllersPath')))
     } catch (e) {
-      logger.error(`${appName} fatal error: could not load controller files from ${app.get('controllersPath')}\n`, e)
+      logger.error(`${appName} fatal error: could not load controller files from ${app.get('controllersPath')}`)
+      logger.error(e)
     }
 
-    // assign controllers to routers
-    if (controllerFiles[0]) {
-      let controllerAssigned
-      let requiredController
-      let routerControllers
-      const mainRouter = app.get('express').Router()
-      controllerFiles.forEach(controller => {
-        controllerAssigned = false
-        if (controller !== app.get('params').errorPages.notFound) {
-          // require controller file
-          const exceptions = ['.DS_Store'] // do not attempt to execute controller logic on these files
-          if (exceptions.indexOf(path.basename(controller)) < 0) {
-            try {
-              requiredController = null
-              if (fs.statSync(controller).isFile() && path.extname(controller) === '.js') {
-                requiredController = require(controller)
-              } else {
-                return
-              }
+    // load all controllers
+    for (let controllerName of controllerFiles) {
+      let controller
+      controllerName = controllerName.path
 
-              // check to see if that controller belongs to a custom router
-              if (routers.controllers) {
-                routers.controllers.forEach((router, routerIndex) => {
-                  // remove duplicates from controllers
-                  routerControllers = [...new Set(router.files)]
-                  if (!routers.controllers[routerIndex].router) {
-                    routers.controllers[routerIndex].router = app.get('express').Router()
-                    // router level middleware to override res.redirect()
-                    routers.controllers[routerIndex].router.use((req, res, next) => {
-                      const redirect = res.redirect
-                      // overwrite the res.redirect() function
-                      res.redirect = function () {
-                        let address
-                        let override
-                        let status
+      if (controllerName !== params.errorPages.notFound) {
+        try {
+          if (fs.statSync(controllerName).isFile() && path.extname(controllerName) === '.js') {
+            controller = require(controllerName)
 
-                        // remove any trailing slashes for proper concatenation
-                        const prefix = router.prefix.replace(/\/+$/, '')
-
-                        // logic to preserve allowable express arguments
-                        if (arguments.length === 1 && typeof arguments[0] === 'string') {
-                          address = arguments[0]
-                        } else if (arguments.length === 2 && typeof arguments[0] === 'number' && typeof arguments[1] === 'string') {
-                          status = arguments[0]
-                          address = arguments[1]
-                        } else if (arguments.length === 2 && typeof arguments[0] === 'string' && typeof arguments[1] === 'boolean') {
-                          address = arguments[0]
-                          override = arguments[1]
-                        } else if (arguments.length === 3 && typeof arguments[0] === 'number' && typeof arguments[1] === 'string' && typeof arguments[2] === 'boolean') {
-                          status = arguments[0]
-                          address = arguments[1]
-                          override = arguments[2]
-                        } else {
-                          // invalid arguments passed to res.redirect(), set address to 'back' so the application doesn't break
-                          logger.error('Invalid arguments supplied to res.redirect()')
-                          address = 'back'
-                        }
-
-                        if (!override && address.charAt(0) === '/') {
-                          address = prefix + address
-                        }
-
-                        if (status) {
-                          redirect.call(this, status, address)
-                        } else {
-                          redirect.call(this, address)
-                        }
-                      }
-                      next()
-                    })
-                  }
-                  routerControllers.forEach(routerController => {
-                    if (controller.includes(path.join(app.get('controllersPath'), routerController))) {
-                      // if the controller accepts less than one or more than two arguments, it's not defining a route
-                      if (requiredController.length > 0 && requiredController.length < 3 && typeof requiredController === 'function') {
-                        requiredController(routers.controllers[routerIndex].router, app)
-                        controllerAssigned = true
-                      }
-                    }
-                  })
-                })
-              }
-              // controller wasn't used so pass it to the main app router
-              if (!controllerAssigned) {
-                // if the controller accepts less than one or more than two arguments, it's not defining a route
-                if (requiredController.length > 0 && requiredController.length < 3 && typeof requiredController === 'function') {
-                  requiredController(mainRouter, app)
-                }
-              }
-            } catch (e) {
-              logger.error(`${appName} failed to load controller file: ${controller}. Please make sure it is coded correctly. See documentation at http://github.com/rooseveltframework/roosevelt for examples.\n`, e)
+            // if the controller accepts less than one or more than two arguments, it's not defining a route
+            if (controller.length > 0 && controller.length < 3) {
+              controller(router, app)
             }
           }
+        } catch (e) {
+          logger.error(`${appName} failed to load controller file: ${controllerName}. Please make sure it is coded correctly. See documentation at http://github.com/rooseveltframework/roosevelt for examples.`)
+          logger.error(e)
         }
-      })
-
-      // have app consume the custom routers
-      if (routers.controllers) {
-        routers.controllers.forEach(router => {
-          if (router.router) {
-            app.use(router.prefix, router.router)
-          }
-        })
       }
-
-      // load 404 controller last so that it doesn't supersede the others
-      try {
-        require(params.errorPages.notFound)(mainRouter, app)
-      } catch (e) {
-        logger.error(`${appName} failed to load 404 controller file: ${params.errorPages.notFound}. Please make sure it is coded correctly. See documentation at http://github.com/rooseveltframework/roosevelt for examples.\n`, e)
-      }
-
-      // have app consume the main router
-      app.use('/', mainRouter)
     }
   }
+
+  // load 404 controller last so that it doesn't supersede the others
+  try {
+    require(params.errorPages.notFound)(router, app)
+  } catch (e) {
+    logger.error(`${appName} failed to load 404 controller file: ${params.errorPages.notFound}. Please make sure it is coded correctly. See documentation at http://github.com/rooseveltframework/roosevelt for examples.`)
+    logger.error(e)
+  }
+
+  // activate the router module
+  app.use(prefix || '/', router)
 
   return app
 }
