+++ lib/scripts/configAuditor.js
@@ -210,8 +210,8 @@ function configAudit (appDir) {
         case 'publicFolder':
           checkTypes(userParam, key, ['string'])
           break
-        case 'routers':
-          checkTypes(userParam, key, ['boolean', 'object'])
+        case 'routePrefix':
+          checkTypes(userParam, key, ['null', 'string'])
           break
         case 'shutdownTimeout':
           checkTypes(userParam, key, ['number'])
