+++ lib/sourceParams.js
@@ -256,8 +256,8 @@ module.exports = (params, app) => {
         default: defaults.errorPages.serviceUnavailable
       }
     },
-    routers: {
-      default: defaults.routers
+    routePrefix: {
+      default: defaults.routePrefix
     },
     staticsRoot: {
       default: defaults.staticsRoot
@@ -465,6 +465,24 @@ module.exports = (params, app) => {
     }
   }
 
+  // sanitize the routePrefix param
+  if (params.routePrefix && typeof params.routePrefix === 'string') {
+    params.routePrefix = params.routePrefix.trim()
+
+    // append a slash if one doesn't exist
+    if (!params.routePrefix.startsWith('/')) {
+      params.routePrefix = `/${params.routePrefix}`
+    }
+
+    // remove trailing slash if one exists
+    if (params.routePrefix.endsWith('/')) {
+      params.routePrefix = params.routePrefix.slice(0, -1)
+    }
+  } else {
+    // default it to empty string if not set
+    params.routePrefix = ''
+  }
+
   // make the app directory requirable
   appModulePath.addPath(appDir)
 
@@ -496,6 +514,9 @@ module.exports = (params, app) => {
 
     // expose app version
     app.set('appVersion', pkg.version)
+
+    // map routePrefix
+    app.set('routePrefix', params.routePrefix)
 
     // map mvc paths
     app.set('modelsPath', path.join(appDir, params.modelsPath))
